pyproject.toml
[project]
name = "unified-etl"
version = "0.1.0"
description = "Unified ETL framework: Ibis transforms on Snowflake + DuckDB"
requires-python = ">=3.11"

[tool.uv.workspace]
members = [
    "packages/transforms",
    "packages/backends",
    "packages/schemas",
    "packages/api",
    "packages/dashboard",
    "packages/nfl",
    "packages/ml",
    "validation",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--import-mode=importlib"
pythonpath = [
    "packages/transforms/src",
    "packages/backends/src",
    "packages/schemas/src",
    "packages/api/src",
    "packages/dashboard/src",
    "packages/nfl/src",
    "packages/ml/src",
    "validation/src",
]

[dependency-groups]
dev = [
    "ibis-framework[duckdb,snowflake]>=9.0.0",
    "snowflake-connector-python[secure-local-storage]>=3.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.0.0",
    "polars>=1.0.0",
    "pyarrow>=14.0.0",
    "pytest",
    "python-dotenv>=1.0.0",
    "nfl-data-py>=0.3.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "joblib>=1.3.0",
    "numpy>=1.26.0",
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.27.0",
]
